Adicionando validacoes nos usuarios

+++ controllers/user.py
@@ -8,8 +8,8 @@ from app.infra.exceptions.user_not_found import UserNotFound
 from app.services.address_service import AddressService
 from app.services.user_service import UserService
 
-from app.infra.forms.login_form import LoginForm
 from app.infra.forms.user_form_register import UserFormRegister
+from app.infra.forms.login_form import LoginForm
 from app.infra.forms.user_form_update import UserFormUpdate
 
 user_service = UserService()
@@ -65,19 +65,12 @@ def signup_action():
     """Action for signup user"""
     try:
         request_form = request.form
-        form = UserFormRegister(request_form)
-        if form.validate():
-            user = user_service.insert_user(request_form)
-            address_service.insert_address(user, request_form)
+        user = user_service.insert_user(request_form)
+        address_service.insert_address(user, request_form)
 
-            return redirect('/listar-usuarios')
-        return redirect(url_for('signup'))
-    except UserExists as e:
-        flash('Usuário já existe')
-        return redirect(url_for('signup'))
-
+        return redirect(url_for('login'))
     except Exception as e:
-        flash('Erro ao cadastrar um novo usuário')
+        flash(str(e))
         return redirect(url_for('signup'))
 
 
@@ -95,10 +88,20 @@ def list_users():
 @is_admin
 def delete_user(user_id: int):
     """Delete an User"""
-    user, address = user_service.get_user_with_address(user_id)
-    address_service.delete(address.id)
-    user_service.delete(user.id)
-    return redirect(url_for('list_users'))
+    try:
+        user = user_service.show(user_id)
+
+        if not user:
+            raise UserNotFound('Usuário não encontrado')
+
+        if user.address[0]:
+            address_service.delete(user.address[0].id)
+        user_service.delete(user.id)
+        return redirect(url_for('list_users'))
+    except Exception as e:
+        flash(str(e))
+        return redirect(url_for('list_users'))
+
 
 
 @app.route('/editar-usuario/<user_id>', methods=['GET'])
@@ -107,11 +110,14 @@ def delete_user(user_id: int):
 def edit_user(user_id: int):
     """Edit an User"""
     try:
-        user, address = user_service.get_user_with_address(user_id)
+        user = user_service.show(user_id)
+        if not user:
+            raise UserNotFound('Usuário não encontrado')
+        
         form = UserFormUpdate()
-        return render_template('signup.html', user=user, address=address, title="Editar Usuário", form=form)
-    except UserNotFound as e:
-        flash('Usuário não encontrado')
+        return render_template('signup.html', user=user, title="Editar Usuário", form=form)
+    except Exception as e:
+        flash(str(e))
         return redirect(url_for('list_users'))
 
 
@@ -120,7 +126,8 @@ def edit_user(user_id: int):
 def edit_profile():
     """Edit profile of current user"""
     form = UserFormUpdate()
-    return render_template('signup.html', title="Editar Perfil", form=form)
+    user = current_user
+    return render_template('signup.html', user=user, title="Editar Perfil", form=form)
 
 
 @app.route('/editar-perfil-action', methods=['POST'])
@@ -141,7 +148,7 @@ def edit_profile_action():
         flash('Campos invalidos')
         return redirect(url_for('edit_profile'))
     except Exception as e:
-        flash('Erro ao atualizar usuário')
+        flash(str(e))
         return redirect(url_for('list_users'))
 
 
@@ -152,15 +159,15 @@ def edit_user_action(user_id):
     """Action for edit user"""
     try:
         request_form = request.form
-        form = UserFormUpdate(request_form)
 
-        if form.validate():
-            address_id = form['address_id']
-            user_service.update_user(user_id, form)
-            address_service.update_address(address_id, form)
-            return redirect(url_for('edit_user', user_id=user_id))
+        user = user_service.update_user(user_id, request_form)
+
+        if user.address[0]:
+            address_service.update_address(user.address[0], request_form)
+
+        return redirect(url_for('edit_user', user_id=user_id))
     except Exception as e:
-        flash('Erro ao atualizar usuário')
+        flash(str(e))
         return redirect(url_for('list_users'))
